make script hierarchical
